+++ sdk/core/core-rest-pipeline/test/defaultRetryPolicy.spec.ts
@@ -54,7 +54,7 @@ describe("defaultRetryPolicy", function () {
     });
     await clock.runAllAsync();
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
   });
 });