+++ sdk/core/core-rest-pipeline/test/retryPolicy.spec.ts
@@ -94,7 +94,7 @@ describe("retryPolicy", function () {
     });
     await clock.runAllAsync();
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
   });
 
@@ -173,7 +173,7 @@ describe("retryPolicy", function () {
     });
     await clock.runAllAsync();
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
     assert.strictEqual(request.url, "https://not-bing.com");
   });
@@ -281,7 +281,7 @@ describe("retryPolicy", function () {
     });
     await clock.runAllAsync();
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
 
     assert.deepEqual(
@@ -294,21 +294,7 @@ describe("retryPolicy", function () {
         "Retry 2: Attempting to send request [Request Id]",
         "Retry 2: Processing 1 retry strategies.",
         "Retry 3: Attempting to send request [Request Id]",
-        "Retry 3: Processing 1 retry strategies.",
-        "Retry 4: Attempting to send request [Request Id]",
-        "Retry 4: Processing 1 retry strategies.",
-        "Retry 5: Attempting to send request [Request Id]",
-        "Retry 5: Processing 1 retry strategies.",
-        "Retry 6: Attempting to send request [Request Id]",
-        "Retry 6: Processing 1 retry strategies.",
-        "Retry 7: Attempting to send request [Request Id]",
-        "Retry 7: Processing 1 retry strategies.",
-        "Retry 8: Attempting to send request [Request Id]",
-        "Retry 8: Processing 1 retry strategies.",
-        "Retry 9: Attempting to send request [Request Id]",
-        "Retry 9: Processing 1 retry strategies.",
-        "Retry 10: Attempting to send request [Request Id]",
-        "Retry 10: Maximum retries reached. Returning the last received response, or throwing the last received error.",
+        "Retry 3: Maximum retries reached. Returning the last received response, or throwing the last received error.",
       ]
     );
 
@@ -319,13 +305,6 @@ describe("retryPolicy", function () {
         "Retry 1: Received an error from request [Request Id]",
         "Retry 2: Received an error from request [Request Id]",
         "Retry 3: Received an error from request [Request Id]",
-        "Retry 4: Received an error from request [Request Id]",
-        "Retry 5: Received an error from request [Request Id]",
-        "Retry 6: Received an error from request [Request Id]",
-        "Retry 7: Received an error from request [Request Id]",
-        "Retry 8: Received an error from request [Request Id]",
-        "Retry 9: Received an error from request [Request Id]",
-        "Retry 10: Received an error from request [Request Id]",
       ]
     );
 
@@ -337,20 +316,6 @@ describe("retryPolicy", function () {
         "Retry 1: Retry strategy testRetryStrategy retries after 100",
         "Retry 2: Processing retry strategy testRetryStrategy.",
         "Retry 2: Retry strategy testRetryStrategy retries after 100",
-        "Retry 3: Processing retry strategy testRetryStrategy.",
-        "Retry 3: Retry strategy testRetryStrategy retries after 100",
-        "Retry 4: Processing retry strategy testRetryStrategy.",
-        "Retry 4: Retry strategy testRetryStrategy retries after 100",
-        "Retry 5: Processing retry strategy testRetryStrategy.",
-        "Retry 5: Retry strategy testRetryStrategy retries after 100",
-        "Retry 6: Processing retry strategy testRetryStrategy.",
-        "Retry 6: Retry strategy testRetryStrategy retries after 100",
-        "Retry 7: Processing retry strategy testRetryStrategy.",
-        "Retry 7: Retry strategy testRetryStrategy retries after 100",
-        "Retry 8: Processing retry strategy testRetryStrategy.",
-        "Retry 8: Retry strategy testRetryStrategy retries after 100",
-        "Retry 9: Processing retry strategy testRetryStrategy.",
-        "Retry 9: Retry strategy testRetryStrategy retries after 100",
       ],
       error: [],
     });
@@ -397,7 +362,7 @@ describe("retryPolicy", function () {
     });
     await clock.runAllAsync();
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
 
     assert.deepEqual(
@@ -410,21 +375,7 @@ describe("retryPolicy", function () {
         "Retry 2: Attempting to send request [Request Id]",
         "Retry 2: Processing 1 retry strategies.",
         "Retry 3: Attempting to send request [Request Id]",
-        "Retry 3: Processing 1 retry strategies.",
-        "Retry 4: Attempting to send request [Request Id]",
-        "Retry 4: Processing 1 retry strategies.",
-        "Retry 5: Attempting to send request [Request Id]",
-        "Retry 5: Processing 1 retry strategies.",
-        "Retry 6: Attempting to send request [Request Id]",
-        "Retry 6: Processing 1 retry strategies.",
-        "Retry 7: Attempting to send request [Request Id]",
-        "Retry 7: Processing 1 retry strategies.",
-        "Retry 8: Attempting to send request [Request Id]",
-        "Retry 8: Processing 1 retry strategies.",
-        "Retry 9: Attempting to send request [Request Id]",
-        "Retry 9: Processing 1 retry strategies.",
-        "Retry 10: Attempting to send request [Request Id]",
-        "Retry 10: Maximum retries reached. Returning the last received response, or throwing the last received error.",
+        "Retry 3: Maximum retries reached. Returning the last received response, or throwing the last received error.",
       ]
     );
 
@@ -435,13 +386,6 @@ describe("retryPolicy", function () {
         "Retry 1: Received an error from request [Request Id]",
         "Retry 2: Received an error from request [Request Id]",
         "Retry 3: Received an error from request [Request Id]",
-        "Retry 4: Received an error from request [Request Id]",
-        "Retry 5: Received an error from request [Request Id]",
-        "Retry 6: Received an error from request [Request Id]",
-        "Retry 7: Received an error from request [Request Id]",
-        "Retry 8: Received an error from request [Request Id]",
-        "Retry 9: Received an error from request [Request Id]",
-        "Retry 10: Received an error from request [Request Id]",
       ]
     );
 
@@ -453,20 +397,6 @@ describe("retryPolicy", function () {
         "Retry 1: Skipped.",
         "Retry 2: Processing retry strategy testRetryStrategy.",
         "Retry 2: Skipped.",
-        "Retry 3: Processing retry strategy testRetryStrategy.",
-        "Retry 3: Skipped.",
-        "Retry 4: Processing retry strategy testRetryStrategy.",
-        "Retry 4: Skipped.",
-        "Retry 5: Processing retry strategy testRetryStrategy.",
-        "Retry 5: Skipped.",
-        "Retry 6: Processing retry strategy testRetryStrategy.",
-        "Retry 6: Skipped.",
-        "Retry 7: Processing retry strategy testRetryStrategy.",
-        "Retry 7: Skipped.",
-        "Retry 8: Processing retry strategy testRetryStrategy.",
-        "Retry 8: Skipped.",
-        "Retry 9: Processing retry strategy testRetryStrategy.",
-        "Retry 9: Skipped.",
       ],
       error: [],
     });
@@ -643,7 +573,7 @@ describe("retryPolicy", function () {
     await clock.runAllAsync();
 
     // should be one more than the default retry count
-    assert.strictEqual(next.callCount, 11);
+    assert.strictEqual(next.callCount, 4);
     assert.isTrue(catchCalled);
     assert.strictEqual(request.url, "https://not-bing.com");
 
@@ -657,21 +587,7 @@ describe("retryPolicy", function () {
         "Retry 2: Attempting to send request [Request Id]",
         "Retry 2: Processing 1 retry strategies.",
         "Retry 3: Attempting to send request [Request Id]",
-        "Retry 3: Processing 1 retry strategies.",
-        "Retry 4: Attempting to send request [Request Id]",
-        "Retry 4: Processing 1 retry strategies.",
-        "Retry 5: Attempting to send request [Request Id]",
-        "Retry 5: Processing 1 retry strategies.",
-        "Retry 6: Attempting to send request [Request Id]",
-        "Retry 6: Processing 1 retry strategies.",
-        "Retry 7: Attempting to send request [Request Id]",
-        "Retry 7: Processing 1 retry strategies.",
-        "Retry 8: Attempting to send request [Request Id]",
-        "Retry 8: Processing 1 retry strategies.",
-        "Retry 9: Attempting to send request [Request Id]",
-        "Retry 9: Processing 1 retry strategies.",
-        "Retry 10: Attempting to send request [Request Id]",
-        "Retry 10: Maximum retries reached. Returning the last received response, or throwing the last received error.",
+        "Retry 3: Maximum retries reached. Returning the last received response, or throwing the last received error.",
       ]
     );
 
@@ -682,13 +598,6 @@ describe("retryPolicy", function () {
         "Retry 1: Received an error from request [Request Id]",
         "Retry 2: Received an error from request [Request Id]",
         "Retry 3: Received an error from request [Request Id]",
-        "Retry 4: Received an error from request [Request Id]",
-        "Retry 5: Received an error from request [Request Id]",
-        "Retry 6: Received an error from request [Request Id]",
-        "Retry 7: Received an error from request [Request Id]",
-        "Retry 8: Received an error from request [Request Id]",
-        "Retry 9: Received an error from request [Request Id]",
-        "Retry 10: Received an error from request [Request Id]",
       ]
     );
 
@@ -700,20 +609,6 @@ describe("retryPolicy", function () {
         "Retry 1: Retry strategy testRetryStrategy redirects to https://not-bing.com",
         "Retry 2: Processing retry strategy testRetryStrategy.",
         "Retry 2: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 3: Processing retry strategy testRetryStrategy.",
-        "Retry 3: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 4: Processing retry strategy testRetryStrategy.",
-        "Retry 4: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 5: Processing retry strategy testRetryStrategy.",
-        "Retry 5: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 6: Processing retry strategy testRetryStrategy.",
-        "Retry 6: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 7: Processing retry strategy testRetryStrategy.",
-        "Retry 7: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 8: Processing retry strategy testRetryStrategy.",
-        "Retry 8: Retry strategy testRetryStrategy redirects to https://not-bing.com",
-        "Retry 9: Processing retry strategy testRetryStrategy.",
-        "Retry 9: Retry strategy testRetryStrategy redirects to https://not-bing.com",
       ],
       error: [],
     });